Move configureEnvironment to run_common.go
